Add the motors package for simulating motors and encoders
Summary:
Adds the motor wrapper structure to simulate motors controlled by TalonFX and SparkMax. All core classes to this PR are mostly in systems/motors, which includes the LoggedMotor interface, and the MotorManager, TalonFXWrapper and SparkMaxWrapper. Values are logged on SmartDashboard. Also added a superclass for FSMs to avoid null pointer errors in main.

* add parent FSMSystem class and modify current structure to use it

* Revert FSMSystem changes

---------

Co-authored-by: AG6GR <AG6GR@users.noreply.github.com>